requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
Flask-Login>=0.6.0
Flask-WTF>=1.1.0
Flask-Mail>=0.9.0
Flask-CORS>=4.0.0
celery>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.0
stripe>=7.0.0
google-cloud-storage>=2.10.0
google-auth>=2.20.0
sentry-sdk[flask]>=1.35.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
gunicorn>=21.0.0
pytest>=7.4.0
pytest-flask>=1.2.0
pytest-cov>=4.1.0
factory-boy>=3.2.0
faker>=19.0.0
qrcode[pil]>=7.4.0 